requirements.txt
Pillow~=9.0.0
scipy~=1.7.0
numpy~=1.22.0
SoundFile~=0.10.0
gradio